requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-CORS==4.0.0
Flask-Limiter==3.3.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
shap==0.42.1
matplotlib==3.7.2
seaborn==0.12.2
reportlab==4.0.9
PyMySQL==1.1.0
redis==4.6.0
python-dotenv==1.0.0
joblib==1.3.1
pydantic==2.1.1
gunicorn==21.2.0
pytest==7.4.0
pytest-cov==4.1.0
openpyxl==3.1.2
xlrd==2.0.1

